+++ src/data_inclusion/schema/models.py
@@ -31,6 +31,7 @@ class Service(BaseModel):
     formulaire_en_ligne: Optional[HttpUrl]
     commune: Optional[str]
     code_postal: Optional[constr(min_length=5, max_length=5, regex=r"^\d{5}$")]
+    code_insee: Optional[constr(min_length=5, max_length=5)]
 
     class Config:
         extra = Extra.forbid