+++ src/data_inclusion/schema/models.py
@@ -33,6 +33,7 @@ class Service(BaseModel):
     code_postal: Optional[constr(min_length=5, max_length=5, regex=r"^\d{5}$")]
     code_insee: Optional[constr(min_length=5, max_length=5)]
     adresse: Optional[str]
+    complement_adresse: Optional[str]
 
     class Config:
         extra = Extra.forbid