+++ src/data_inclusion/schema/modes_accueil.py
@@ -1,0 +1,14 @@
+from data_inclusion.schema.base import EnhancedEnum
+
+
+class ModeAccueil(EnhancedEnum):
+    A_DISTANCE = (
+        "a-distance",
+        "À distance",
+        None,
+    )
+    EN_PRESENTIEL = (
+        "en-presentiel",
+        "En présentiel",
+        None,
+    )