+++ src/data_inclusion/schema/code_officiel_geographique.py
@@ -1,0 +1,19 @@
+from data_inclusion.schema.base import EnhancedEnum
+
+
+class TypeCOG(EnhancedEnum):
+    COMMUNE = (
+        "commune",
+        "Commune",
+        None,
+    )
+    REGION = (
+        "region",
+        "Région",
+        None,
+    )
+    DEPARTEMENT = (
+        "departement",
+        "Département",
+        None,
+    )