+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * `presentation_detail`
     * `pre_requis`
     * `cumulable`
+    * `justificatifs`
 
 ## Version 0.7.0 - 2022-12-08