+++ CHANGELOG.md
@@ -14,6 +14,7 @@
     * `formulaire_en_ligne`
     * `commune`
     * `code_postal`
+    * `code_insee`
 
 ## Version 0.7.0 - 2022-12-08