+++ CHANGELOG.md
@@ -20,6 +20,7 @@
     * `longitude`
     * `latitude`
     * `recurrence`
+    * `date_creation`
 
 ## Version 0.7.0 - 2022-12-08