+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     * `latitude`
     * `recurrence`
     * `date_creation`
+    * `date_suspension`
 
 ## Version 0.7.0 - 2022-12-08