+++ CHANGELOG.md
@@ -22,6 +22,7 @@
     * `recurrence`
     * `date_creation`
     * `date_suspension`
+    * `lien_source`
 
 ## Version 0.7.0 - 2022-12-08