+++ CHANGELOG.md
@@ -23,6 +23,7 @@
     * `date_creation`
     * `date_suspension`
     * `lien_source`
+    * `telephone`
 
 ## Version 0.7.0 - 2022-12-08