+++ CHANGELOG.md
@@ -25,6 +25,7 @@
     * `lien_source`
     * `telephone`
     * `courriel`
+    * `contact_public`
 
 ## Version 0.7.0 - 2022-12-08