+++ CHANGELOG.md
@@ -26,6 +26,7 @@
     * `telephone`
     * `courriel`
     * `contact_public`
+    * `date_maj`
 
 ## Version 0.7.0 - 2022-12-08